\name{CWMix-package}
\alias{CWMix-package}
\alias{CWMix}
\docType{package}
\title{
\packageTitle{CWMix}
}
\description{
\packageDescription{CWMix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CWMix}
\packageIndices{CWMix}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CWMix}

Maintainer: \packageMaintainer{CWMix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
